style(ui): add navbar layout without functionality
