+++ frontend/src/components/navigation/Navbar.tsx
@@ -2,20 +2,25 @@ import DarkLightSwitch from "./DarkLightSwitch";
 import HomeButton from "./HomeButton";
 import ProfileButton from "./ProfileButton";
 import SearchBar from "./SearchBar";
+import { RoundedButton } from "../ui/Buttons";
 
 export default function Navbar() {
   return (
-    <nav className="fixed dark:bg-black dark:text-white px-8 top-0 w-screen z-50 bg-white h-15 shadow-md flex flex-row justify-between items-center">
-      <div className="flex-1 flex justify-start">
-        <HomeButton />
-      </div>
-      <div className="flex-1 flex shrink-0">
-      <SearchBar />
-      </div>
-      <div className="flex-1 justify-end flex flex-row items-center gap-4">
-        <DarkLightSwitch />
-        <ProfileButton />
-      </div>
-    </nav>
+    <div className="fixed top-0 w-screen px-2">
+      <nav className="backdrop-blur-sm mx-auto z-50 bg-white/20 border-white/30 border-1 shadow-md flex flex-row justify-between items-center rounded-full mt-2 p-2 px-4">
+        <div className="flex-1 flex justify-start">
+          <RoundedButton>
+            <HomeButton />
+          </RoundedButton>
+        </div>
+        <div className="flex-1 flex shrink-0">
+          <SearchBar />
+        </div>
+        <div className="flex-1 justify-end flex flex-row items-center gap-4">
+          <DarkLightSwitch />
+          <ProfileButton />
+        </div>
+      </nav>
+    </div>
   );
 }
