+++ frontend/src/components/navigation/ProfileButton.tsx
@@ -1,3 +1,14 @@
+"use client"
+import { RoundedButton } from "../ui/Buttons";
+import { useState } from "react";
+import Image from "next/image";
+import { User } from "lucide-react";
 export default function ProfileButton() {
-    return <button>Profile</button>
+    const [isLoggedIn, setIsLoggedIn] = useState(false);
+    const [loggedInUser, setLoggedInUser] = useState(null);
+  return (
+    <RoundedButton>
+      {isLoggedIn ? <Image src={loggedInUser.profilePicture} width={64} height={64} alt="Profile Picture"/> : <User/>}
+    </RoundedButton>
+  );
 }
